requirements.txt
discord.py==2.3.1
firebase_admin==6.2.0
gTTS==2.3.2
Pillow==10.0.0
Pillow==11.0.0
python-dotenv==1.0.1
rich==13.9.3
scrapetube==2.5.1
